fix(builder): fixed fluent api for the required rule builder
The required rule builder was in need of 2 optional parameters: path and
errorMessage.
However errorMessage can be specified regardless whether
path is present or not, making the api cumbersome.
Changed the builder
adding the withPath, withErrorMessage and build methods

BREAKING CHANGE: Now the required rule builder requires the build method
to be called